refactor: Isolate all functions for drawing from main
