+++ Makefile
@@ -4,7 +4,7 @@ CXXFLAGS = -std=c++17 -g -Wall -Wextra -Wno-sign-compare -Wno-unused-parameter
 #LDFLAGS = -L/usr/X11R6/lib -L/usr/pkg/lib
 LDLIBS  = -lglut -lGLU -lGL -lm
 
-$(PROGRAM): main.o image.o mole.o sargarepa.o rotkvica.o pecurka.o coveculjak.o
+$(PROGRAM): main.o image.o mole.o sargarepa.o rotkvica.o pecurka.o coveculjak.o iscrtavanja.o
 	$(CXX) -o $@ $(CXXFLAGS) $^ $(LDLIBS)
 
 image.o: image.c image.h
@@ -26,6 +26,9 @@ pecurka.o: pecurka.cpp pecurka.hpp
 	$(CXX) -c -o $@ $(CXXFLAGS) $< $(LDLIBS)
 
 coveculjak.o: coveculjak.cpp coveculjak.hpp
+	$(CXX) -c -o $@ $(CXXFLAGS) $< $(LDLIBS)
+
+iscrtavanja.o: iscrtavanja.cpp iscrtavanja.hpp
 	$(CXX) -c -o $@ $(CXXFLAGS) $< $(LDLIBS)
 
 .PHONY: clean run
